reverted finished code back to original after modified by rohit

+++ Finished_Code_HershJ/car/receiver.cpp
@@ -23,11 +23,11 @@ void receiver_setup() {
   radio2.printDetails();
 }
 
-bool receiving(int &val) {
+bool receiving(float &val) {
     // if there is data ready
     bool result;
     if ( radio2.available() ) {
-      result = radio2.read(&val, sizeof(int));
+      result = radio2.read(&val, sizeof(float));
       //printf(" %d \n", val);
       return result;
     }
